SearchBar 위젯으로 분리, 스타일 변경
